Put array answers in the right place, added async and flowControl
Modified some of the array answers to use filter instead of manually
iterating.  I stuck with manual iteration for "count", just because I
figured it was more efficient than doing a filter

The async answers are really awkward.  I don't know if it was intended
to be a test of library familiarity.  I just looked at the test,
glanced at the Promises/A spec, and wrote something to make it pass.